Fix NaN/Inf issues in ROME optimization for Qwen3.5 models
- Add gradient clipping and NaN detection in compute_v.py
- Use float32 for delta optimization with proper dtype casting
- Cap learning rate at 0.01 for stability
- Track best delta and restore on NaN
- Use bfloat16 instead of float16 in mtloader.py
- Add Colab tutorial notebook in examples/

+++ fastedit/utils/mtloader.py
@@ -30,7 +30,7 @@ def load_model_and_tokenizer(
     model = AutoModelForCausalLM.from_pretrained(
         model,
         low_cpu_mem_usage=True,
-        torch_dtype=torch.float16,
+        torch_dtype=torch.bfloat16,  # Use bfloat16 for better numerical stability
         trust_remote_code=True
     ).cuda()
